chore: remove jest-fetch-mock dependency
