+++ package.json
@@ -61,7 +61,6 @@
     "eslint": "^8.53.0",
     "eslint-config-prettier": "^9.0.0",
     "jest": "^29.3.1",
-    "jest-fetch-mock": "^3.0.3",
     "jsdoc": "^4.0.2",
     "minami": "^1.2.3",
     "prettier": "^3.0.3",
@@ -69,8 +68,7 @@
     "ts-mockito": "^2.6.1",
     "ts-node": "^10.9.1",
     "typedoc": "^0.25.3",
-    "typescript": "^5.2.2",
-    "whatwg-fetch": "^3.6.19"
+    "typescript": "^5.2.2"
   },
   "dependencies": {
     "@noble/hashes": "^1.3.2",
